ref #30 BIG refactoring
