+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "submodules/QuasarAppLib"]
 	path = submodules/QuasarAppLib
 	url = https://github.com/QuasarApp/QuasarAppLib
+[submodule "submodules/SimpleQmlNotify"]
+	path = submodules/SimpleQmlNotify
+	url = https://github.com/QuasarApp/SimpleQmlNotify.git
