+++ CMakeLists.txt
@@ -45,6 +45,9 @@ endif()
 
 # Add sub directories
 
+add_subdirectory(submodules/QuasarAppLib)
+add_subdirectory(submodules/SimpleQmlNotify)
+
 add_subdirectory(src/ClientLib)
 add_subdirectory(src/Client)
