+++ src/Client/main.cpp
@@ -1,10 +1,20 @@
-//#
-//# Copyright (C) 2021-2021 QuasarApp.
-//# Distributed under the lgplv3 software license, see the accompanying
-//# Everyone is permitted to copy and distribute verbatim copies
-//# of this license document, but changing it is not allowed.
-//#
+#include <QGuiApplication>
+#include <QQmlApplicationEngine>
+#include <QQmlContext>
+#include "clientapp.h"
 
-int main(int argc, char *argv[]) {
-    return 0;
+
+int main(int argc, char *argv[])
+{
+    QCoreApplication::setAttribute(Qt::AA_EnableHighDpiScaling);
+
+    QGuiApplication app(argc, argv);
+    QQmlApplicationEngine engine;
+    ClientApp client;
+
+    if (!client.init(&engine)) {
+        return 1;
+    }
+
+    return app.exec();
 }
