+++ src/ClientLib/SnakeProject/Private/networkclient.h
