+++ src/ClientLib/SnakeProject/Private/settingsviewmodel.cpp
