+++ src/ClientLib/SnakeProject/Private/settingsviewmodel.h
