--- src/ClientLib/SnakeProject/ProfileViewItems/playerclientdata.cpp
@@ -1,6 +1,0 @@
-#include "playerclientdata.h"
-
-PlayerClientData::PlayerClientData()
-{
-
-}
