--- src/ClientLib/SnakeProject/ProfileViewItems/playerclientdata.h
@@ -1,11 +1,0 @@
-#ifndef PLAYERCLIENTDATA_H
-#define PLAYERCLIENTDATA_H
-#include <player.h>
-
-class PlayerClientData: public ClientProtocol::Player
-{
-public:
-    PlayerClientData();
-};
-
-#endif // PLAYERCLIENTDATA_H
