+++ src/ClientLib/SnakeProject/snakeprojectsettings.cpp
